How to use for FactorialCalculator.java
The comment is there to guide you on how to customize the program for your specific use case by changing the value of number to the desired input number for the factorial calculation.

For example, if you want to calculate the factorial of 10, you can change it to int number = 10;, and the program will then calculate the factorial of 10. You can modify the value of the number variable to calculate the factorial of any non-negative integer you choose.